# frontend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Trippy API",
    "version": "v1"
  },
  "paths": {
...
    "/api/itineraries": {
      "get": {
        "tags": [
          "itineraries"
        ],
        "operationId": "list_itineraries",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItinerarySummary"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItinerarySummary"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItinerarySummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "itineraries"
        ],
        "operationId": "create_itinerary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryCreate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryCreate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryDetail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/itineraries/{itineraryId}": {
      "get": {
        "tags": [
          "itineraries"
        ],
        "operationId": "get_itinerary",
        "parameters": [
          {
            "name": "itineraryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryDetail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "itineraries"
        ],
        "operationId": "update_itinerary",
        "parameters": [
          {
            "name": "itineraryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryDetail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "itineraries"
        ],
        "operationId": "delete_itinerary",
        "parameters": [
          {
            "name": "itineraryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/places": {
      "get": {
        "tags": [
          "places"
        ],
        "operationId": "list_places",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceRead"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceRead"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceRead"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/places/{placeId}": {
      "get": {
        "tags": [
          "places"
        ],
        "operationId": "get_place",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceRead"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceRead"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/sections": {
      "get": {
        "tags": [
          "sections"
        ],
        "operationId": "list_sections",
        "parameters": [
          {
            "name": "itinerary_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionRead"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionRead"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionRead"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sections"
        ],
        "operationId": "create_section",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionCreate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionCreate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SectionRead"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionRead"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/sections/{sectionId}": {
      "patch": {
        "tags": [
          "sections"
        ],
        "operationId": "update_section",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SectionRead"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionRead"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sections"
        ],
        "operationId": "delete_section",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ItemRead": {
        "required": [
          "description",
          "id",
          "place_id",
          "section_id",
          "sequence"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "section_id": {
            "type": "string",
            "format": "uuid"
          },
          "place_id": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "sequence": {
            "type": "integer",
            "format": "int32"
          },
          "start_time": {
            "type": "string",
            "nullable": true
          },
          "place": {
            "$ref": "#/components/schemas/PlaceRead"
          }
        },
        "additionalProperties": false
      },
...
      "ItineraryCreate": {
        "required": [
          "name",
          "start_date"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start_date": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "ItineraryDetail": {
        "required": [
          "description",
          "end_date",
          "id",
          "name",
          "start_date"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "start_date": {
            "type": "string",
            "format": "date"
          },
          "end_date": {
            "type": "string",
            "format": "date"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionRead"
            }
          }
        },
        "additionalProperties": false
      },
      "ItinerarySummary": {
        "required": [
          "description",
          "end_date",
          "id",
          "name",
          "start_date"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "start_date": {
            "type": "string",
            "format": "date"
          },
          "end_date": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "ItineraryUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "start_date": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PlaceRead": {
        "required": [
          "city",
          "description",
          "id",
          "latitude",
          "longitude",
          "name",
          "region",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "minLength": 1,
            "type": "string"
          },
          "city": {
            "minLength": 1,
            "type": "string"
          },
          "region": {
            "minLength": 1,
            "type": "string"
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "hours": {
            "type": "string",
            "nullable": true
          },
          "duration_minutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "price_range": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "seasonal_notes": {
            "type": "string",
            "nullable": true
          },
          "booking_required": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "SectionCreate": {
        "required": [
          "date",
          "itinerary_id"
        ],
        "type": "object",
        "properties": {
          "itinerary_id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "sequence": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SectionRead": {
        "required": [
          "date",
          "description",
          "id",
          "itinerary_id",
          "sequence"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "itinerary_id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "sequence": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemRead"
            }
          }
        },
        "additionalProperties": false
      },
      "SectionUpdate": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "sequence": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}